VideoPlay.jsx page updated
